Pass text to and from C++ library via Horos
I have built the first part of the program: a C++ controller object
that is wrapped in Objective C. It’s not much at this point but it
accomplished a really important task which was passing a NSString
object to the C++ controller, which then spit it back to Horos. Not
much as far as fireworks, but it means I can successfully interface a
C++ program with the Objective C application (API).