Secondary Creation + Ability to read the current DICOM image
This is a secondary creation of the project due to an issue with the
previous version not being seen in Horos. It is a reversion of sorts
that also has added functionality.

It can now load the current image as an NSArray of image slices in the
NSArray object pointer *Pixarray = [[viewerController imageView]
dcmPixList]

The number of slices is then returned to the user